Fix search function & implement destructuring
This commit fixes the search functionality so that the results shown on
the page are up to date with the search term in the search box. Prior
to this commit the search results were one keystroke behind the search
term due to the asynchronous nature of useState. Adding in the
useEffect hook forces the re-render to occur AFTER the search term has
been updated so the filter on allSpells is checking against the most
current search term.

This commit also adds in destructuring to the props when accessed in
child components so they can be accessed directly. Now accessing
spellName is done as spell.spellName instead of spell.spell.spellName.
This doesn't affect functionality but does look cleaner.