Add Buttons to Remove Spells
This commit adds in remove functionality(in beta form). There is now a
'Remove Spell' button inside each spell component that will remove that
spell from the 'spellsToDisplay' array. This functionality will need to be
redone after we implement individual spellbooks since the remove will only
remove the spell from the spellbook and not from the 'allSpells' array.

Since we don't want to actually remove a spell from the 'allSpells' array,
in the current version of the application the spell is displayed again if
a spell is searched for and the removed spell meets the search criteria or
if the search criteria is cleared out. Once individual spellbooks are
implemented then the remove button will actually remove the spell from
that spellbook's array of spells so this won't be an issue at that point.
Removing spells will only be a feature inside of spellbooks and not
available inside the 'allSpells' container(whatever it is called then).