Add version 1 of filtering spells
This is the beginning of adding filters to the application. This first
iteration works but can be buggy when multiple filters are checked
simultaneously. Espcially if the filters are of the same type(school,
level, etc). Future versions will allow for multiple filters to be applied
at once to display that meet any of the filters. This way users can search
for spells by multiple schools, levels, etc. Eventually the filters will
be able to be applied to be 'or' filters(as described above) or they can
be 'and' filters to find spells that meet all of the selected criteria.