Upgrade filters to allow multiple filters at once The previous code allowed some filtering but would not allow for multiple filters to be used simultaneously in the future. The changes in this commit allow for much more flexibility in filtering in future versions of the app.
This commit allows for multiple filters to be applied but it only really
works if the filters are of the same type(such as 'school' or 'level').
If filters of different types are selected the most recent one will
override the others. Ex: Selecting the 'Abjuration' and 'Conjuration'
filters will disply only abjuration and conjuration spells. If 'Level 1'
is then selected all level 1 spells will be displayed(not just level 1
abjuration or conjuratinspells). More complicated filtering logic will
 be required for that functionality. 'Or' filtering
will have to only apply to filters of the same type while 'and' filtering
will be used across multiple types(Selecting 'Abjuration', 'Conjuration',
and 'Level 1' should display spells that are level 1 and either abjuration or
conjuration instead of showing level 1 of all spell schools.